Time: 5 ms (39.85%), Space: 57.1 MB (6.93%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -7,10 +7,8 @@ class Solution {
         for(j=0;j<k;j++)
         {
             sum+=nums[j];
-            
-         
         }
-        avg = (double) (sum/k);
+        avg = (sum/k);
         for(j=k;j<nums.length;j++)
         {
           
@@ -18,7 +16,7 @@ class Solution {
             i++;
             sum+=nums[j];
           
-            avg= Math.max(avg,(double)(sum/k));
+            avg= Math.max(avg,(sum/k));
         }
         return avg;
